Q3G7HM8 consolidate lib/scrubber to five modules: layout, axis, series, time, hooks
